+++ CHANGELOG.rst
@@ -36,6 +36,7 @@ Changed
 - added functions module for stateless transforms
 - removed narwhals private import from nominal module `#725 <https://github.com/azukds/tubular/issues/725>_`
 - setup get_transform_exprs methods for aggregations/comparison/dates/misc/numeric, and OneHotEncoder/MeanResponseTransformer
+- bug: fixed MeanResponseTransformer.fit to accept both lazy X and lazy y `#731 <https://github.com/azukds/tubular/pull/731>`_
 
 3.3.0 (30/04/2026)
 ------------------